requirements.txt
Flask
pdfplumber
sentence-transformers
faiss-cpu